+++ lib/src/generators/provider_generator.dart
@@ -54,7 +54,7 @@ class ProviderGenerator {
     final featurePath = p.join(projectPath, 'lib', 'features', snakeFeature);
     if (!FileUtils.directoryExistsSync(featurePath)) {
       ConsoleUtils.error('Feature "$featureName" does not exist.');
-      ConsoleUtils.info('Run "fcli g feature $featureName" first.');
+      ConsoleUtils.info('Run "flg g feature $featureName" first.');
       return;
     }