+++ lib/src/generators/widget_generator.dart
@@ -67,7 +67,7 @@ class WidgetGenerator {
     final featurePath = p.join(projectPath, 'lib', 'features', snakeFeature);
     if (!FileUtils.directoryExistsSync(featurePath)) {
       ConsoleUtils.error('Feature "$featureName" does not exist.');
-      ConsoleUtils.info('Run "fcli g feature $featureName" first.');
+      ConsoleUtils.info('Run "flg g feature $featureName" first.');
       return;
     }