+++ test/integration/init_test.dart
@@ -1,9 +1,9 @@
 import 'dart:io';
 
-import 'package:fcli/src/config/config_loader.dart';
-import 'package:fcli/src/config/fcli_config.dart';
-import 'package:fcli/src/generators/project_generator.dart';
-import 'package:fcli/src/utils/file_utils.dart';
+import 'package:flg/src/config/config_loader.dart';
+import 'package:flg/src/config/fcli_config.dart';
+import 'package:flg/src/generators/project_generator.dart';
+import 'package:flg/src/utils/file_utils.dart';
 import 'package:test/test.dart';
 
 void main() {